refactor: 결제를 위한 계산 로직 분리 및 적용
- getOrderCalculationData 함수 생성
